backend/requirements.txt
fastapi
uvicorn[standard]
sqlmodel
psycopg2-binary
requests
beautifulsoup4
python-dotenv
nltk
google-generativeai
playwright